Move code to framework project.
